backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.1
factory-boy==2.12.0
Faker==4.0.3
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.3
pathspec==0.8.0
psycopg2-binary==2.8.5
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.4.4
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
